requirements.txt
python-telegram-bot==20.7
mysql-connector-python==8.3.0 